+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Script for running local environment with Kind and Helm
 * Creating Kubernetes templates for Helm
 * Creation of develop branch for leeting "main" as production branch
 * Initial commit from local